+++ model/src/main/java/model/element/Position.java
@@ -1,27 +1,27 @@
 package model.element;
 
 public class Position implements IPosition {
-	private float x;
-	private float y;
+	private int x;
+	private int y;
 	
-	public Position(float x,float y) {
+	public Position(int x,int y) {
 		this.setX(x);
 		this.setY(y);
 	}
 	
-	public float getX() {
+	public int getX() {
 		return x;
 	}
 	
-	public void setX(float x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 	
-	public float getY() {
+	public int getY() {
 		return y;
 	}
 	
-	public void setY(float y) {
+	public void setY(int y) {
 		this.y = y;
 	}
 }
